requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-telegram-bot[fast]==20.7
langchain==0.1.20
langchain-google-vertexai==0.0.6
langchain-community==0.0.36
faiss-cpu==1.7.4
PyPDF2==3.0.1